Don't directly inline contexts
While I don't see how a malicious actor would be able to influence
these contexts I'd rather be in the habbit of always passing contexts
to scripts by way of environment variables. Also, I suspect that it
this way is more likely that an unexpected context will result in a
meaningful error message.